src/requirements.txt
asn1crypto==0.24.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.2
datapunt-authorization-django==0.2.17
datapunt-authorization-levels==0.1.3
Django==1.11
django-braces==1.10.0
django-extensions==1.7.4
django-filter==1.0.2
django-rest-swagger==2.1.0
djangorestframework==3.6.2
djangorestframework-gis==0.10.1
drf-extensions==0.3.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.6.1
pycparser==2.18
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
